+++ app/controllers/support/establishments_controller.rb
@@ -1,0 +1,11 @@
+module Support
+  class EstablishmentsController < ApplicationController
+    def index
+      respond_to do |format|
+        format.json do
+          render json: EstablishmentSearch.omnisearch(params[:q]).as_json
+        end
+      end
+    end
+  end
+end