+++ app/forms/support/create_case_form_schema.rb
@@ -6,12 +6,14 @@ module Support
     config.messages.top_namespace = :case_migration_form
 
     params do
-      required(:school_urn).value(:string)
       required(:first_name).value(:string)
       required(:last_name).value(:string)
       required(:email).value(:string)
       optional(:request_type).value(:bool)
       optional(:organisation_id).value(:string)
+      optional(:organisation_type).value(:string)
+      optional(:organisation_name).value(:string)
+      optional(:organisation_urn).value(:string)
       optional(:phone_number).value(:string)
       optional(:category_id).value(:string)
       optional(:hub_case_ref).value(:string)
@@ -27,26 +29,25 @@ module Support
     # TODO: custom macro using chronic
     rule(:estimated_procurement_completion_date).validate(format?: /(^$|^\d{2}\/\d{2}\/\d{4}$)/)
 
-    rule(:school_urn) do
-      key(:school_urn).failure(:missing) if value.blank?
-      key(:school_urn).failure(:invalid) if invalid_school_urn?(value)
-    end
-
     # TODO: add email validation format
     rule(:email) do
       key(:email).failure(:missing) if value.blank?
+    end
+
+    rule(:organisation_id) do
+      key(:organisation_id).failure(:missing) if value.blank?
     end
 
     rule(:category_id) do
       key.failure(:missing) if values[:request_type].presence && value.blank?
     end
 
-  private
+    rule(:organisation_type) do
+      key(:organisation_id).failure(:missing) if value.blank?
+    end
 
-    def invalid_school_urn?(value)
-      return true if value.blank?
-
-      Support::Organisation.find_by(urn: value).nil?
+    rule(:organisation_name) do
+      key(:organisation_id).failure(:missing) if value.blank?
     end
   end
 end