+++ app/presenters/support/create_case_form_presenter.rb
@@ -1,7 +1,11 @@
 # frozen_string_literal: true
 
+require_relative "concerns/has_organisation"
+
 module Support
   class CreateCaseFormPresenter < BasePresenter
+    include Support::Concerns::HasOrganisation
+
     # @return [String]
     def full_name
       "#{first_name} #{last_name}"
@@ -12,11 +16,6 @@ module Support
       return I18n.t("support.case_categorisations.label.none") if category_id.blank?
 
       @category_name ||= Support::Category.find(category_id).title
-    end
-
-    # @return [Support::Organisation]
-    def establishment
-      @establishment ||= Support::Organisation.find_by(urn: school_urn)
     end
   end
 end