+++ db/migrate/20220204135146_add_group_uid_to_framework_requests.rb
@@ -1,0 +1,5 @@
+class AddGroupUidToFrameworkRequests < ActiveRecord::Migration[6.1]
+  def change
+    add_column :framework_requests, :group_uid, :string
+  end
+end