+++ spec/factories/framework_requests.rb
@@ -2,6 +2,7 @@ FactoryBot.define do
   factory :framework_request do
     first_name { "David" }
     last_name { "Georgiou" }
+    group { false }
     school_urn { "000001" }
     email { "email@example.com" }
     message_body { "please help!" }