+++ spec/features/self-serve/faf/submit_framework_request_spec.rb
@@ -16,6 +16,7 @@ RSpec.feature "Completed framework requests" do
   end
 
   before do
+    create(:support_organisation, urn: "100253", name: "School #1")
     stub_request(:post, "https://api.notifications.service.gov.uk/v2/notifications/email")
       .with(body: email.to_json)
       .to_return(body: {}.to_json, status: 200, headers: {})
@@ -30,7 +31,7 @@ RSpec.feature "Completed framework requests" do
     click_on "Send request"
   end
 
-  xit "shows the confirmation page" do
+  it "shows the confirmation page" do
     expect(find("h1.govuk-panel__title")).to have_text "Your request for support has been sent"
     expect(find("div.govuk-panel__body")).to have_text "We have sent a confirmation email to: #{user.email}"